:electron: Add new publication
Add the article on seismicity along the Baribis fault, Indonesia.

Widiyantoro, S., Supendi, P., Ardianto, A. et al. Implications for fault
locking south of Jakarta from an investigation of seismic activity along
the Baribis fault, northwestern Java, Indonesia. Sci Rep 12, 10143 (2022).
https://doi.org/10.1038/s41598-022-13896-6

I used QuakeMigrate to generate an earthquake catalogue for the region
based on the available stations. With this, I was able to expand the
existing catalogue and further strengthen the inferences and conclusions
of the initial manuscript.