:electron: Add new publication
Add the article on deep long period seismicity on the Reykjanes
Peninsula, 2021.

Greenfield, T., Winder, T., Rawlinson, N. et al. Deep long period seismicity
preceding and during the 2021 Fagradalsfjall eruption, Iceland. Bull Volcanol
84, 101 (2022). https://doi.org/10.1007/s00445-022-01603-2

I contributed to the fieldwork and development of the software, QuakeMigrate,
that was key for detecting the deep long period earthquakes studied in
this paper, as well as commenting on/editing the manuscript.